databricks/content/dbx_app_with_pg/requirements.txt
agate
aiofiles
altair
annotated-types
anyio
appdirs
asgiref
attrs
babel
backports-tarfile
blinker
cachetools
certifi
charset-normalizer
click
cloudpickle
colorama
contourpy
cycler
daff
dash
dash-ag-grid
dash-bootstrap-components
dash-core-components
dash-html-components
dash-mantine-components
dash-table
databricks-sdk==0.57.0
databricks-sql-connector
dbt-adapters
dbt-common
dbt-core
dbt-databricks
dbt-extractor
dbt-protos
dbt-semantic-interfaces
dbt-spark
debugpy
deepdiff
et-xmlfile
fastapi
ffmpy
filelock
flask
fonttools
fsspec
gitdb
gitpython
google-auth
gradio
gradio-client
gunicorn
h11
hf-xet
htmltools
httpcore
httptools
httpx
huggingface-hub
idna
importlib-metadata
importlib-resources
isodate
itsdangerous
jaraco-classes
jaraco-context
jaraco-functools
jinja2
jsonschema
jsonschema-specifications
keyring
kiwisolver
leather
linkify-it-py
lz4
markdown-it-py
markupsafe
mashumaro
matplotlib
mdit-py-plugins
mdurl
mlflow-skinny
more-itertools
msgpack
narwhals
nest-asyncio
networkx
numpy
oauthlib
openpyxl
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
orjson
packaging
pandas
parsedatetime
pathspec
pillow
plotly
plotly-resampler
prompt-toolkit
protobuf
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydeck
pydub
pygments
pyparsing
python-dateutil
python-dotenv
python-multipart
python-slugify
pytimeparse
pytz
pyyaml
questionary
referencing
requests
retrying
rich
rpds-py
rsa
ruff
semantic-version
setuptools
shellingham
shiny
six
smmap
sniffio
snowplow-tracker
sqlparams
sqlparse
starlette
streamlit
tenacity
text-unidecode
thrift
toml
tomlkit
tornado
tqdm
tsdownsample
typer
typing-extensions
typing-inspection
tzdata
uc-micro-py
urllib3
uvicorn
uvloop
watchfiles
wcwidth
websockets
werkzeug
zipp
sqlalchemy
